+++ src/components/common/navigationGuard/useGuardedNavigate.ts
